Add bearer authorization check for clearing user sessions

+++ src/app/api/sessions/clear/route.ts
@@ -1,4 +1,4 @@
-import { NextResponse } from "next/server"
+import { NextRequest, NextResponse } from "next/server"
 
 // Lib
 import { db } from "@lib/database.lib"
@@ -6,7 +6,24 @@ import { db } from "@lib/database.lib"
 // Utils
 import { ServiceError } from "@utils/helpers/error.helpers"
 
-export async function DELETE() {
+// Setup
+const { CRON_SECRET = "" } = process.env
+
+/**
+ * Normally this should be a DELETE method.
+ *
+ * However, we mostly trigger it as a cron job via Vercel
+ * where the only supported verb is GET.
+ */
+export async function GET(req: NextRequest) {
+  const authHeader = req.headers.get("authorization")
+
+  if (authHeader !== `Bearer ${CRON_SECRET}`) {
+    return NextResponse.json(new ServiceError("Unauthorized"), {
+      status: 401,
+    })
+  }
+
   try {
     await db.deleteFrom("Session").where("expires", "<", new Date()).execute()
